data-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3
openpyxl==3.1.2
GitPython==3.1.40
pydantic-settings==2.1.0

